Minor tweaks
no functional change

+++ PKHeX.Core/Legality/Checks.cs
@@ -32,6 +32,6 @@ namespace PKHeX.Core
 
         public static string[] MoveStrings { internal get; set; } = Util.GetMovesList("en");
         public static string[] SpeciesStrings { internal get; set; } = Util.GetSpeciesList("en");
-        internal static IEnumerable<string> GetMoveNames(IEnumerable<int> moves) => moves.Select(m => m >= MoveStrings.Length ? L_AError : MoveStrings[m]);
+        internal static IEnumerable<string> GetMoveNames(IEnumerable<int> moves) => moves.Select(m => (uint)m >= MoveStrings.Length ? L_AError : MoveStrings[m]);
     }
 }
